+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 
 0.2.4 (YYYY-MM-DD)
 ------------------
+* Work around casacore getcolslice caching (:pr:`107`)
 * Update LICENSE year (:pr:`105`)
 * Update license and production status in pypi classifiers (:pr:`104`)
 * Use WHERE rather than HAVING clause in group ordering TAQL (:pr:`98`)